Expand vocabulary from 30K to 100K words (float16, 57MB)
- 3x more words: covers everyday words like 포옹, 장모, etc.
- float16 storage halves file size (57MB vs 115MB for float32)
- float32 at runtime for computation accuracy

+++ setup_vectors.py
@@ -140,7 +140,7 @@ def _process_gz_file(gz_path, max_words, vectors_path, words_path):
 
 
 def _save(words, vectors, dim, vectors_path, words_path):
-    vectors_np = np.array(vectors, dtype=np.float32)
+    vectors_np = np.array(vectors, dtype=np.float16)
     np.save(vectors_path, vectors_np)
     with open(words_path, "w", encoding="utf-8") as f:
         json.dump(words, f, ensure_ascii=False)
